Update driver.py print message
beginner users might be confused by this printed message. Regardless of whether the process successfully work or not, it prints same message with log file path. 

But when process doesn't work because of several each user's env reason, the error log just printed on the {job_name}_logging under root directory. So in this case they need to check this file.